Added QBox2DBody and revised folder arrangements

+++ qbox2dcommon.h
@@ -1,6 +1,6 @@
 #ifndef QBOX2DCOMMON_H
 #define QBOX2DCOMMON_H
 
-static int sizeMultiplier = 30; //use this as a size factor for items in QBox2DWorld
+static float sizeMultiplier = 30; //use this as a size factor for items in QBox2DWorld
 
 #endif // QBOX2DCOMMON_H
